requirements.txt
Django==1.7.4
django-filter==0.9.2
djangorestframework==3.0.4
flup==1.0.2
wsgiref==0.1.2
drf-extensions==0.2.7
